Surface build trial

+++ CoonsSurfaceConstructor/convexhull.cpp
@@ -18,11 +18,11 @@ void ConvexHull::build() {
 }
 
 Point ConvexHull::getMinXPoint() {
-    float minX = 100000;
+    double minX = 100000;
     Point resPoint;
     QVector<Point>::iterator i;
     for (i = begin(); i != end(); ++i) {
-        minX = std::min(i->x(), minX);
+        minX = std::min((double)i->x(), minX);
         if(i->x() == minX) {
             if(!(resPoint.x() < i->x() && resPoint.x() == minX))
                 resPoint = *i;
